:sparkles: feat(AnimalApi, Swagger) cria pasta para animal api e implementa dependencia para acessar swagger
