+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
 spring.application.name=safe-pet-backend
-spring.data.mongodb.uri={}
+spring.data.mongodb.uri=mongodb+srv://devbruno:root@cluster0.xvjnwqv.mongodb.net/dbinfo?retryWrites=true&w=majority&appName=Cluster0
